Huge redesign
This redesign greatly simplifies both the required plugin options, as
well as the required dependencies.  It makes use of a new
`buildkite-bash-tools` repository for greater confidence when doing
things like treehashing, and it also does a few crazy things like
constructing signed AWS requests in pure `openssl`/`bash`.  This has the
extreme benefit of working on our build machines without any hint of the
`aws` executable however, although it does require the `zstd` executable
because nobody has time to wait around for `gzip` to compress things.